+++ BO360/BO360/BOPowerUp.cpp
@@ -37,15 +37,15 @@ void BOPowerUp::Shutdown()
 
 void BOPowerUp::Draw()
 {
-	BOGraphicInterface::DrawEx(m_sprite, int4(0,0,40,40), int4(m_position.x,m_position.y,m_size.x,m_size.y), 0.0, int2(0,0));
+	BOGraphicInterface::DrawEx(m_sprite, int4(0, 0, 40, 40), int4((int)m_position.x, (int)m_position.y, m_size.x, m_size.y), 0.0, int2(0, 0));
 }
 
 
 void BOPowerUp::Update(double p_deltaTime)
 {
 	// Move the powerup towards the blackhole
-	m_position.x = (m_speed * p_deltaTime) * m_direction.x + m_position.x;
-	m_position.y = (m_speed * p_deltaTime) * m_direction.y + m_position.y;
+	m_position.x = (float)(m_speed * p_deltaTime) * m_direction.x + m_position.x;
+	m_position.y = (float)(m_speed * p_deltaTime) * m_direction.y + m_position.y;
 
 	// Checks powerup "ball" against the pad, if colliding with pad do powerup effect and remove the powerup"ball"
 	float2 result = BOPhysics::BallPadCollision(GetBoundingSphere(), m_direction, m_paddle->GetBoundingSphere(), m_paddle->GetRotation() - 10.6, m_paddle->GetDegrees());
